demos/customer-service/requirements.txt
openai-agents
chainlit
pydantic
python-dotenv
